+++ src/Silverback.Integration/Messaging/Producing/DefaultProduceStrategy.cs
@@ -2,6 +2,9 @@
 // This code is licensed under MIT license (see LICENSE file for details)
 
 using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Threading;
 using System.Threading.Tasks;
 using Silverback.Messaging.Configuration;
 using Silverback.Messaging.Messages;
@@ -46,6 +49,86 @@ public sealed class DefaultProduceStrategy : IProduceStrategy, IEquatable<Defaul
             Check.NotNull(envelope, nameof(envelope));
 
             await envelope.Producer.ProduceAsync(envelope).ConfigureAwait(false);
+        }
+
+        [SuppressMessage("ReSharper", "AccessToModifiedClosure", Justification = "Intentional and not an issue here.")]
+        [SuppressMessage("Maintainability", "CA1508:Avoid dead conditional code", Justification = "False positive.")]
+        public async Task ProduceAsync(IEnumerable<IOutboundEnvelope> envelopes)
+        {
+            Check.NotNull(envelopes, nameof(envelopes));
+
+            // The extra 1 is to prevent the completion before the iteration is over and all messages are enqueued
+            int pending = 1;
+            TaskCompletionSource<bool> taskCompletionSource = new();
+            Exception? produceException = null;
+
+            foreach (IOutboundEnvelope envelope in envelopes)
+            {
+                if (taskCompletionSource.Task.IsCompleted)
+                    break;
+
+                Interlocked.Increment(ref pending);
+
+                envelope.Producer.Produce(
+                    envelope,
+                    _ =>
+                    {
+                        if (Interlocked.Decrement(ref pending) == 0)
+                            taskCompletionSource.TrySetResult(true);
+                    },
+                    exception =>
+                    {
+                        produceException = exception;
+
+                        if (Interlocked.Decrement(ref pending) == 0)
+                            taskCompletionSource.TrySetResult(false);
+                    });
+            }
+
+            if (Interlocked.Decrement(ref pending) > 0)
+                await taskCompletionSource.Task.ConfigureAwait(false);
+
+            if (produceException != null)
+                throw produceException;
+        }
+
+        [SuppressMessage("ReSharper", "AccessToModifiedClosure", Justification = "Intentional and not an issue here.")]
+        [SuppressMessage("Maintainability", "CA1508:Avoid dead conditional code", Justification = "False positive.")]
+        public async Task ProduceAsync(IAsyncEnumerable<IOutboundEnvelope> envelopes)
+        {
+            // The extra 1 is to prevent the completion before the iteration is over and all messages are enqueued
+            int pending = 1;
+            TaskCompletionSource<bool> taskCompletionSource = new();
+            Exception? produceException = null;
+
+            await foreach (IOutboundEnvelope envelope in envelopes)
+            {
+                if (taskCompletionSource.Task.IsCompleted)
+                    break;
+
+                Interlocked.Increment(ref pending);
+
+                envelope.Producer.Produce(
+                    envelope,
+                    _ =>
+                    {
+                        if (Interlocked.Decrement(ref pending) == 0)
+                            taskCompletionSource.TrySetResult(true);
+                    },
+                    exception =>
+                    {
+                        produceException = exception;
+
+                        if (Interlocked.Decrement(ref pending) == 0)
+                            taskCompletionSource.TrySetResult(false);
+                    });
+            }
+
+            if (Interlocked.Decrement(ref pending) > 0)
+                await taskCompletionSource.Task.ConfigureAwait(false);
+
+            if (produceException != null)
+                throw produceException;
         }
     }
 }